Add script for quick installation of k3s
To allow for quick and easy installation of k3s, a new script called `quick_install_k3s.sh` has been added. This script assumes that k3s will be running on a single node, where the master node and worker nodes are on the same server.

The script uses `curl` to download the k3s installation script and then runs it using `sh`. It also creates the necessary directory for the kubeconfig file and generates the kubeconfig file using the `/etc/rancher/k3s/k3s.yaml` file. Finally, it provides instructions for adding an environment variable to the `.shrc` file to set the `KUBECONFIG` variable.

Changes:
- Added `quick_install_k3s.sh` script for quick installation of k3s
- Generated kubeconfig file using `/etc/rancher/k3s/k3s.yaml`
- Provided instructions for adding environment variable to `.shrc` file