Create nginx-pod.yaml file
In order to deploy an nginx web server, a new file called nginx-pod.yaml was created. This file contains the necessary configuration to create a Kubernetes Pod with an nginx container. The Pod is named "my-nginx-pod" and it uses the "nginx:latest" image. The container listens on port 80 using the TCP protocol.